requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
dblpy==0.3.4
discord.py==1.3.3
emoji==0.5.4
fuzzywuzzy==0.18.0
hijri-converter==2.1.1
html2text==2020.1.16
html5lib==1.1
idna==2.10
multidict==4.7.6
psutil==5.7.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.13.0
python-Levenshtein==0.12.0
ratelimiter==1.2.0.post0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
webencodings==0.5.1
websockets==8.1
yarl==1.4.2
youtube-dl==2020.6.16.1
